get functions and templates
the get label and description functions now return template values in the specific case where a template is set but a label/description is not. This is done to make print functions more clear when someone has set a template for a variable but hasnt yet formalized that template into something specific.